Refactor offline queue
This changes the offline queue to use a database
(independent from the one for saving links) instead of abusing the
sharedPreferences for that and adds a button to the 3-dot-menu to
manually save queued links.
Additionally, this fixes some errors and adds more checks that should
avoid queueing a link multiple times.

Concerns #11